requirements.txt
python-dotenv>=1.0.0
pandas>=2.0.0
pymysql>=1.1.0
openpyxl>=3.1.0
google>=3.0.0
scrapegraphai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0 